Update rakefile example app to 8.0
