+++ examples/rakefile-activerecord8.0/Gemfile
@@ -1,6 +1,6 @@
 source 'https://rubygems.org'
 
-gem 'activerecord', '7.0.1', require: 'active_record'
+gem 'activerecord', '8.0.1', require: 'active_record'
 gem 'otr-activerecord', path: '../../'
 gem 'sqlite3'
 gem 'rake'
