+++ examples/rakefile-activerecord8.0/Rakefile
