+++ examples/rakefile-activerecord8.0/data/migrate/20160731142314_create_widgets.rb
