+++ examples/rakefile-activerecord8.0/data/migrate/20171128231939_create_splines.rb
