+++ examples/rakefile-activerecord8.0/data/schema.rb
