+++ examples/rakefile-activerecord8.0/data/seeds.rb
